fix: imports, moved translation to page title component
