+++ src/components/common/Extensions/MainNavigation/partials/hooks.ts
@@ -1,4 +1,5 @@
 import { useMemo } from 'react';
+
 import { NavItem } from './types';
 
 export const useGetNavItems = (name: string | undefined): NavItem[] =>
