+++ src/components/frame/v5/pages/TeamsPage/TeamsPage.tsx
@@ -1,6 +1,7 @@
 import React, { FC } from 'react';
-import Navigation from '~v5/common/Navigation/Navigation';
-import TwoColumns from '~v5/frame/TwoColumns/TwoColumns';
+
+import Navigation from '~v5/common/Navigation';
+import TwoColumns from '~v5/frame/TwoColumns';
 
 const TeamsPage: FC = () => (
   <TwoColumns aside={<Navigation pageName="members" />}>Teams page</TwoColumns>
